Fixed warnings

+++ src/object/path_walker.cpp
@@ -143,7 +143,7 @@ PathWalker::jump_to_node(int node_no)
   Path* path = get_path();
   if (!path) return;
 
-  if (node_no >= path->get_nodes().size()) return;
+  if (node_no >= static_cast<int>(path->get_nodes().size())) return;
   m_next_node_nr = static_cast<size_t>(node_no);
   if (m_walking_speed > 0) {
     advance_node();
